Tag endpoints return the helpful error messages (#189)
* Tag endpoints return the helpful error messages

* Update backend/src/main/java/ai/verta/modeldb/versioning/RepositoryDAORdbImpl.java

+++ backend/src/main/java/ai/verta/modeldb/versioning/RepositoryDAORdbImpl.java
@@ -293,9 +293,23 @@ public class RepositoryDAORdbImpl implements RepositoryDAO {
       session.beginTransaction();
       RepositoryEntity repository = getRepositoryById(session, request.getRepositoryId());
 
-      // TODO: Check commit_hash exists in DB
-      TagsEntity tagsEntity =
-          new TagsEntity(repository.getId(), request.getCommitSha(), request.getTag());
+      boolean exists =
+          VersioningUtils.commitRepositoryMappingExists(
+              session, request.getCommitSha(), repository.getId());
+      if (!exists) {
+        throw new ModelDBException(
+            "Commit_hash and repository_id mapping not found", Code.NOT_FOUND);
+      }
+
+      Query query = session.createQuery(GET_TAG_HQL);
+      query.setParameter("repositoryId", repository.getId());
+      query.setParameter("tag", request.getTag());
+      TagsEntity tagsEntity = (TagsEntity) query.uniqueResult();
+      if (tagsEntity != null) {
+        throw new ModelDBException("Tag '" + request.getTag() + "' already exists", Code.NOT_FOUND);
+      }
+
+      tagsEntity = new TagsEntity(repository.getId(), request.getCommitSha(), request.getTag());
       session.save(tagsEntity);
       session.getTransaction().commit();
       return SetTagRequest.Response.newBuilder().build();
@@ -312,6 +326,10 @@ public class RepositoryDAORdbImpl implements RepositoryDAO {
       query.setParameter("repositoryId", repository.getId());
       query.setParameter("tag", request.getTag());
       TagsEntity tagsEntity = (TagsEntity) query.uniqueResult();
+      if (tagsEntity == null) {
+        throw new ModelDBException("Tag not found", Code.NOT_FOUND);
+      }
+
       CommitEntity commitEntity = session.get(CommitEntity.class, tagsEntity.getCommit_hash());
       session.getTransaction().commit();
       return GetTagRequest.Response.newBuilder().setCommit(commitEntity.toCommitProto()).build();
@@ -324,8 +342,10 @@ public class RepositoryDAORdbImpl implements RepositoryDAO {
       session.beginTransaction();
       RepositoryEntity repository = getRepositoryById(session, request.getRepositoryId());
       TagsEntity tagsEntity =
-          session.load(
-              TagsEntity.class, new TagsEntity.TagId(request.getTag(), repository.getId()));
+          session.get(TagsEntity.class, new TagsEntity.TagId(request.getTag(), repository.getId()));
+      if (tagsEntity == null) {
+        throw new ModelDBException("Tag not found", Code.NOT_FOUND);
+      }
       session.delete(tagsEntity);
       session.getTransaction().commit();
       return DeleteTagRequest.Response.newBuilder().build();
@@ -341,6 +361,10 @@ public class RepositoryDAORdbImpl implements RepositoryDAO {
       Query query = session.createQuery(GET_TAGS_HQL);
       query.setParameter("repoId", repository.getId());
       List<TagsEntity> tagsEntities = query.list();
+
+      if (tagsEntities == null || tagsEntities.isEmpty()) {
+        return ListTagsRequest.Response.newBuilder().setTotalRecords(0).build();
+      }
 
       session.getTransaction().commit();
       List<String> tags =
